Converts to float, removes plotting calls for profiling

+++ include/io_util.h
@@ -4,6 +4,6 @@
 #include <string>
 class io_util{
     public:
-        static std::vector<double> import_vector_csv(const std::string& path);
-        static std::vector<std::vector<double>> import_matrix_csv(const std::string& path);
+        static std::vector<float> import_vector_csv(const std::string& path);
+        static std::vector<std::vector<float>> import_matrix_csv(const std::string& path);
 };
